refactor: 경로 알맞게 이동
