+++ src/main/java/com/bitharmony/comma/member/auth/dto/GoogleMemberResponse.java
@@ -1,4 +1,4 @@
-package com.bitharmony.comma.member.member.dto;
+package com.bitharmony.comma.member.auth.dto;
 
 import com.bitharmony.comma.member.member.entity.Member;
