+++ src/main/java/com/bitharmony/comma/member/auth/provider/GithubAuthProvider.java
@@ -1,8 +1,8 @@
-package com.bitharmony.comma.global.provider;
+package com.bitharmony.comma.member.auth.provider;
 
 import com.bitharmony.comma.global.exception.member.GetTokenFailureException;
-import com.bitharmony.comma.member.member.dto.GithubMemberResponse;
-import com.bitharmony.comma.member.member.dto.GithubOauthResponse;
+import com.bitharmony.comma.member.auth.dto.GithubMemberResponse;
+import com.bitharmony.comma.member.auth.dto.GithubOauthResponse;
 import java.util.HashMap;
 import java.util.Map;
 import lombok.Getter;
